third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-event-listeners-expected.txt

Test that event listeners not user reachable from the root are still present in the class list.

PASS: the class name is found
